add @Overrides annotations and nuke unnecessary casts
git-svn-id: https://wala.svn.sourceforge.net/svnroot/wala/trunk@1198 f5eafffb-2e1d-0410-98e4-8ec43c5233c4

+++ com.ibm.wala.shrike/src/com/ibm/wala/shrikeCT/ConstantPoolParser.java
@@ -413,7 +413,7 @@ public final class ConstantPoolParser implements ClassConstants {
   }
 
   private long getLong(int i) {
-    return ((long) getInt(i) << 32) + ((long) getInt(i + 4) & 0xFFFFFFFFL);
+    return ((long) getInt(i) << 32) + (getInt(i + 4) & 0xFFFFFFFFL);
   }
 
   private float getFloat(int i) {
